Added google auth gem
