+++ app/controllers/users/omniauth_callbacks_controller.rb
@@ -11,21 +11,4 @@ class Users::OmniauthCallbacksController < Devise::OmniauthCallbacksController
           redirect_to new_user_registration_url, alert: @user.errors.full_messages.join("\n")
         end
     end
-
-    def self.from_omniauth(access_token)
-        data = access_token.info
-        user = User.where(email: data['email']).first
-    
-        # Uncomment the section below if you want users to be created if they don't exist
-        unless user
-            user = User.create(name: data['name'],
-               first_name: data['first_name'],
-               last_name: data['last_name'],
-               email: data['email'],
-               photo: data['image'],
-               password: Devise.friendly_token[0,20]
-            )
-        end
-        user
-    end
   end
