+++ app/javascript/controllers/index.js
@@ -4,8 +4,8 @@
 
 import { application } from "./application"
 
-import Clipboard from 'stimulus-clipboard'
-application.register('clipboard', Clipboard)
-
 import HelloController from "./hello_controller"
 application.register("hello", HelloController)
+
+import TabController from "./tab_controller"
+application.register("tab", TabController)
