+++ app/javascript/controllers/tab_controller.js
@@ -1,0 +1,16 @@
+import { Controller } from "@hotwired/stimulus"
+
+// Connects to data-controller="tab"
+export default class extends Controller {
+  targets = ["first", "second"]  
+  connect() {
+  }
+
+  // Connects to data-action="click->tab#toggle"
+  toggle(event) {
+    console.log(this.firstTarget)
+    event.preventDefault()
+    this.firstTarget.classList.toggle("hidden")
+    this.secondTarget.classList.toggle("hidden")
+  }
+}
